Made EXWM do multi-monitor support the way I want it to
So we already had every monitor added to EXWM's output list.
This commit creates a workspace for every monitor and gives us nice
bindings to switch between 'workspaces' but it effectively is switching
monitors. The workspaces start from 0, but my bindings start from
1 (because my keyboard is laid out such that 1 is very far from 0).
This means the binding s-1 actually goes to workspace 0 and the
binding s-2 goes to workspace 1.